Update 0.0.5
New keys mechanic.
New light sources mechanic.
New location: bookshop.
New items.
New sprites.
Travelling system rework.

+++ Assets/Scripts/Scriptable objects/LocationData.cs
@@ -6,11 +6,11 @@ public class LocationData : ScriptableObject
 {
     [SerializeField] private string _name;
     [SerializeField, TextArea] private string _description;
-    [SerializeField] private int _distance;
+    [SerializeField] private float _distance;
     [SerializeField, Tooltip("ID in scene manager.")] private string _locationID;
 
     public string Name => _name;
     public string Description => _description;
-    public int Distance => _distance;
+    public float Distance => _distance;
     public string LocationID => _locationID;
 }
